Change C-style casts to static_cast and remove unnecessary semicolons (#668)
* Change C-style casts to static_cast

* Remove unnecessary semicolons

* Fix format

* Use .f


* Use .f


* Remove extra )

* Fix remaining issues

* Fix format

* Disable -Wold-style-cast for MurmurHash3.cpp

* Don't change the flags

* Fix test

---------

+++ tests/root_io/read_and_write_associated.cpp
@@ -24,7 +24,7 @@ void writeCollection() {
     item1.Number(i);
     info.push_back(item1);
 
-    event.putParameter("UserEventWeight", (float)100. * i);
+    event.putParameter("UserEventWeight", 100.f * i);
     std::cout << " event number: " << i << std::endl;
     event.putParameter("UserEventName", std::to_string(i));
